(PC-23190)[API] chore: extract post dates test
